Remove unused region config
Regions.getCurrentRegion() only works when run on EC2 environments.
When off EC2 (eg. running locally or in Lambda) it returns `null`,
and since a recent upgrade to the SDK, leaves a warning and stack
trace in the logs. Since no part of snapshotter relies on the region
config, delete it to avoid the warning and any future confusion.